Fix basic styling
I fixed the list page to display properly. I changed the Link to only work when clicking the Note Title cause there was an issue were the click event would fire on dragdrop and I couldn't find any useful solution to the problem. Now anywhere but the title will allow the user to drag the note. Having trouble sorting a bug that stacks all the notes on top of each other when the app initializes and a few styling problems when dealing with the actively dragging item.