Fix bugs
I fixed a few issues that was preventing click events from triggering on cards in the list. I had to put in a distance delay for registering the drag using the dragStartPredicate option from muuri. Also fixed another issue where a newly created note would be rendered to the list incorrectly.