Refine styling
I fixed a few styling issues that were bothering me. The delete button on each card is now an x box. When the user selects notes to delete, a trash can appears to confirm the deletion. There's an issue with clicking on the delete x when a note is not returned in the user search. The back of the card does not display correctly. More tinkering is needed to figure out what is going wrong there.